+++ internal/model/app.go
@@ -7,8 +7,16 @@ import (
 
 // AppConfig is the configuration of the app.
 type AppConfig struct {
+	// Fs is the Fs configuration.
+	Fs FsConfig
 	// Group configuration by ID
 	Groups map[string]GroupConfig
+}
+
+// FsConfig is the Fs configuration.
+type FsConfig struct {
+	Exclude []string
+	Include []string
 }
 
 // GroupConfig is the group configuration.